Bump to 2.2.2

+++ resources/recipes/prepare_system.rb
@@ -65,13 +65,15 @@ sysmem_total = (node['memory']['total'].to_i * 0.90).to_i
 memory_services(sysmem_total)
 
 # Build service list for rbcli
-services = node['redborder']['services'] || []
-systemd_services = node['redborder']['systemdservices'] || []
-service_enablement = {}
+unless File.exist?('/etc/redborder/services.json')
+  services = node['redborder']['services'] || []
+  systemd_services = node['redborder']['systemdservices'] || []
+  service_enablement = {}
 
-systemd_services.each do |service_name, systemd_name|
-  service_enablement[systemd_name.first] = services[service_name]
+  systemd_services.each do |service_name, systemd_name|
+    service_enablement[systemd_name.first] = services[service_name]
+  end
+
+  Chef::Log.info('Saving services enablement into /etc/redborder/services.json')
+  File.write('/etc/redborder/services.json', JSON.pretty_generate(service_enablement))
 end
-
-Chef::Log.info('Saving services enablement into /etc/redborder/services.json')
-File.write('/etc/redborder/services.json', JSON.pretty_generate(service_enablement))
